[feat](aa wallet): visualize aa wallet investment
