+++ pages/dashboard.tsx
@@ -13,6 +13,7 @@ import { useState, useEffect } from "react";
 import RebalanceChart from "./views/RebalanceChart";
 import { useAccount } from "wagmi";
 import TokenDropdownInput from "./views/TokenDropdownInput.jsx";
+import { AllWeatherPortfolio } from "../classes/AllWeatherPortfolio.js";
 
 interface Pools {
   key: string;
@@ -396,6 +397,14 @@ const Dashboard: NextPage = () => {
         </>
       </div>
       <Space>
+        <RebalanceChart
+          rebalanceSuggestions={[]}
+          netWorth={100}
+          windowWidth={200}
+          showCategory={true}
+          mode="portfolioStrategy"
+          portfolioComposition={Object.entries(AllWeatherPortfolio.strategy)}
+        />
         <TokenDropdownInput
           address={walletAddress}
           onClickCallback={async (
